requirements.txt
openai
python-dotenv
requests
secure-smtplib
fastapi
uvicorn
requests-toolbelt
python-dotenv
py-algorand-sdk
opencv-python